+++ examples/functions_test.dart
@@ -4,27 +4,22 @@ import 'package:adapters_flutter/adapters_flutter.dart';
 
 void main() {
   group('functions', () {
-    tmsTest(
-        'add attachment - success',
-        externalId: 'add_attachment_success',
+    tmsTest('add attachment - success',
         () async => await addAttachment('./examples/attachments/file1.txt'));
 
-    tmsTest('add attachment - failed', externalId: 'add_attachment_failed',
-        () async {
+    tmsTest('add attachment - failed', () async {
       await addAttachment('./examples/attachments/file2.json');
       expect(0, 1);
     });
 
     tmsTest(
         'add attachments - success',
-        externalId: 'add_attachments_success',
         () async => await addAttachments([
               './examples/attachments/file1.txt',
               './examples/attachments/file2.json'
             ]));
 
-    tmsTest('add attachments - failed', externalId: 'add_attachments_failed',
-        () async {
+    tmsTest('add attachments - failed', () async {
       await addAttachments([
         './examples/attachments/file1.txt',
         './examples/attachments/file2.json'
@@ -32,32 +27,25 @@ void main() {
       expect(0, 1);
     });
 
-    tmsTest(
-        'add link - success',
-        externalId: 'add_link_success',
+    tmsTest('add link - success',
         () async => await addLink('https://www.example.org/'));
 
-    tmsTest('add link - failed', externalId: 'add_link_failed', () async {
+    tmsTest('add link - failed', () async {
       await addLink('https://www.example.org/');
       expect(0, 1);
     });
 
-    tmsTest(
-        'add links - success',
-        externalId: 'add_links_success',
+    tmsTest('add links - success',
         () async => await addLinks([Link('https://www.example.org/')]));
 
-    tmsTest('add links - failed', externalId: 'add_links_failed', () async {
+    tmsTest('add links - failed', () async {
       await addLinks([Link('https://www.example.org/')]);
       expect(0, 1);
     });
 
-    tmsTest(
-        'add message - success',
-        externalId: 'add_message_success',
-        () async => await addMessage('message'));
+    tmsTest('add message - success', () async => await addMessage('message'));
 
-    tmsTest('add message - failed', externalId: 'add_message_failed', () async {
+    tmsTest('add message - failed', () async {
       await addMessage('message');
       expect(0, 1);
     });